# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kisara API",
    "description": "API documentation untuk layanan kisara.my.id (generated otomatis dari route schemas).",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/auth/google/url": {
      "get": {
        "summary": "Generate Google OAuth URL",
        "tags": [
          "Auth"
        ],
        "description": "Returns Google OAuth consent URL that will be used by frontend login flow.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "url"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "post": {
        "summary": "Google OAuth callback",
        "tags": [
          "Auth"
        ],
        "description": "Exchange authorization code with Google and issue JWT access/refresh tokens for web session auth.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "minLength": 1,
                    "maxLength": 150,
                    "type": "string"
                  },
                  "state": {
                    "minLength": 1,
                    "maxLength": 150,
                    "type": "string"
                  }
                },
                "required": [
                  "code",
                  "state"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "refresh_token": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "access_token",
                        "refresh_token"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "errors",
                    "metadata"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Refresh JWT token",
        "tags": [
          "Auth"
        ],
        "description": "Validate refresh token and return newly generated access/refresh token pair.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "refresh_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "refresh_token": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "access_token",
                        "refresh_token"
                      ]
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "data",
                    "metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {}
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "errors",
                    "metadata"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/api-keys/{api_key_id}": {
      "delete": {
        "summary": "Revoke API key",
        "tags": [
          "Auth"
        ],
        "description": "Revoke an API key owned by the authenticated user.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "api_key_id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number",
                      "enum": [
                        200
                      ]
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status_code": {
                      "type": "number"
                    },
                    "message": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "properties": {
                        "method": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string"
                        },
                        "execution_time_ms": {
                          "type": "number"
                        },
                        "ip": {
                          "type": "string"
                        },
                        "timestamp": {
                          "format": "date-time",
                          "type": "string"
                        }
                      },
                      "required": [
                        "method",
                        "url",
                        "execution_time_ms",
                        "ip",
                        "timestamp"
                      ]
                    }
                  },
                  "required": [
                    "status_code",
                    "message",
                    "metadata"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.kisara.my.id",
      "description": "Production"
    }
  ],
...
}